Add .withUnsafe method to all query builders that appends an arbitrary URL argument.
This is a future-proofing hack that's needed because I don't have the time to properly fix the
generation for availabilities. Availability is the only argument that has an overload, i.e.,
it can take either a string or an ISO8601 duration, but The Creator chose to just comment
out the duration part and assume we'd never use it. This is not the first time we need a way
to force weird arguments for Nitro calls, so this should also help us in the future.